requirements.txt
Django==1.6.5
argparse==1.2.1
gunicorn==18.0
psycopg2==2.5.3
wsgiref==0.1.2
pyserial